Fixed integration tests and gulp tasks
Fixed Dot e2e test

Serialized webdriver update in e2e tests

Added latest chrome to travis

Added chrome no-sandbox flag to protractor
