+++ package.json
@@ -23,7 +23,10 @@
     "gulp-karma": "0.0.4",
     "gulp-load-plugins": "^0.10.0",
     "gulp-mocha": "^2.1.0",
+    "gulp-nunjucks-render": "^0.2.1",
     "gulp-protractor": "^1.0.0",
+    "gulp-rename": "^1.2.2",
+    "gulp-serve": "^1.0.0",
     "gulp-shell": "^0.4.1",
     "gulp-tslint": "^3.0.1-beta",
     "gulp-typescript": "^2.7.5",
