--- test/e2e/protractor.config.js
@@ -1,4 +1,0 @@
-exports.config = {
-  directConnect: true,
-  specs: ['test_utils.js', 'test_cases/**/spec.js']
-}
