--- test/e2e/scripts/all.html.tmpl
@@ -1,54 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="en" ng-app="example">
-<head>
-  <title>n3-charts</title>
-  <style type="text/css">
-  .example {
-    width: 600px;
-    height: 300px;
-    border: 1px solid lightgrey;
-    margin: 10px auto;
-    padding: 10px;
-  }
-
-  .example .label {
-    font-family: monospace;
-    font-weight: 100;
-  }
-
-  </style>
-</head>
-
-<body ng-controller='ExamplesCtrl'>
-  <div ng-repeat="example in examples" class="example">
-    <div class="label">{{example.label}}</div>
-    <linechart datasets="example.datasets" options="example.options"></linechart>
-  </div>
-
-  <script src="%project_path%node_modules/angular/angular.js"></script>
-  <script src="%project_path%node_modules/d3/d3.min.js"></script>
-  <script src="%project_path%build/LineChart.js"></script>
-  <script type="text/javascript">
-  angular.module('example', ['n3-line-chart'])
-  .controller('ExamplesCtrl', function($scope) {
-    $scope.examples = [
-      %repeat%{datasets: %datasets%, options: %options%, label: "%name%"}%repeat%
-    ];
-
-    setInterval(function() {
-      $scope.examples.forEach(function(example) {
-        example.datasets.dataset0.forEach(function(datum) {
-          datum.val_0 = Math.random()*5;
-          datum.val_1 = Math.random()*5;
-        });
-
-        example.datasets = angular.copy(example.datasets);
-      });
-
-      $scope.$apply();
-
-    }, 1000);
-  });
-  </script>
-</body>
-</html>
