--- test/e2e/scripts/case_template/.gitignore
@@ -1,1 +1,0 @@
-index.html
