--- test/e2e/scripts/each.html.tmpl
@@ -1,35 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="en" ng-app="example">
-<head>
-  <title>n3-charts</title>
-  <style type="text/css">
-  .example {
-    width: 600px;
-    height: 300px;
-    border: 1px solid lightgrey;
-    margin: 10px auto;
-    padding: 10px;
-  }
-
-  .example .label {
-    font-family: monospace;
-    font-weight: 100;
-  }
-
-  </style>
-</head>
-
-<body ng-controller='ExamplesCtrl'>
-  <linechart datasets="example.datasets" options="example.options"></linechart>
-
-  <script src="%project_path%node_modules/angular/angular.js"></script>
-  <script src="%project_path%node_modules/d3/d3.min.js"></script>
-  <script src="%project_path%build/LineChart.js"></script>
-  <script type="text/javascript">
-  angular.module('example', ['n3-line-chart'])
-  .controller('ExamplesCtrl', function($scope) {
-    $scope.example = {datasets: %datasets%, options: %options%, label: "%name%"};
-  });
-  </script>
-</body>
-</html>
