+++ test/e2e/templates/_base.tpl.html
@@ -1,0 +1,39 @@
+<!DOCTYPE HTML>
+<html lang="en" ng-app="example">
+<head>
+  <title>n3-charts</title>
+  <style type="text/css">
+  .example {
+    width: 600px;
+    height: 300px;
+    border: 1px solid lightgrey;
+    margin: 10px auto;
+    padding: 10px;
+  }
+
+  .example .label {
+    font-family: monospace;
+    font-weight: 100;
+  }
+
+  </style>
+</head>
+
+<body ng-controller='ExamplesCtrl'>
+  <linechart datasets="example.datasets" options="example.options"></linechart>
+
+  <script src="/angular/angular.js"></script>
+  <script src="/d3/d3.min.js"></script>
+  <script src="/build/LineChart.js"></script>
+  <script type="text/javascript">
+  angular.module('example', ['n3-line-chart'])
+  .controller('ExamplesCtrl', function($scope) {
+    $scope.example = {
+      datasets: {% block datasets %}{}{% endblock %}, 
+      options: {% block options %}{}{% endblock %}, 
+      label: "{% block name %}{% endblock %}"
+    };
+  });
+  </script>
+</body>
+</html>
