+++ test/e2e/test.e2e.ts
@@ -1,0 +1,4 @@
+/// <reference path='../../src/app.ts' />
+
+/// <reference path='../../typings/angular-protractor/angular-protractor.d.ts' />
+/// <reference path='../../typings/jasmine/jasmine.d.ts' />
