--- test/e2e/test_cases/area_series/.gitignore
@@ -1,1 +1,0 @@
-index.html
