--- test/e2e/test_cases/area_series/options
@@ -1,23 +1,0 @@
-{series: [
-        {
-          axis: "y",
-          dataset: "dataset0",
-          key: "val_0",
-          label: "An area series",
-          color: "#1f77b4",
-          type: ["area", "line", "dot"],
-          id: 'mySeries0'
-        },
-        {
-          axis: "y",
-          dataset: "dataset0",
-          key: "val_1",
-          label: "Another area series",
-          color: "grey",
-          thickness: "1px",
-          type: "area",
-          id: 'mySeries1'
-        }
-      ],
-      axes: {x: {key: "x"}}
-}
