--- test/e2e/test_cases/dot_series/.gitignore
@@ -1,1 +1,0 @@
-index.html
