--- test/e2e/test_cases/dot_series/options
@@ -1,24 +1,0 @@
-{
-  series: [
-    {
-      axis: "y",
-      dataset: "dataset0",
-      key: "val_0",
-      label: "A first series",
-      color: "#1f77b4",
-      type: ["dot", "line"],
-      id: 'mySeries0'
-    },
-    {
-      axis: "y",
-      dataset: "dataset0",
-      key: "val_1",
-      label: "Another series",
-      color: "grey",
-      thickness: "1px",
-      type: "dot",
-      id: 'mySeries1'
-    }
-  ],
-  axes: {x: {key: "x"}}
-}
