--- test/e2e/test_cases/line_series/.gitignore
@@ -1,1 +1,0 @@
-index.html
