--- test/e2e/test_cases/line_series/options
@@ -1,23 +1,0 @@
-{series: [
-        {
-          axis: "y",
-          dataset: "dataset0",
-          key: "val_0",
-          label: "A line sinusoid",
-          color: "#1f77b4",
-          type: "line",
-          id: 'mySeries0'
-        },
-        {
-          axis: "y",
-          dataset: "dataset0",
-          key: "val_1",
-          label: "Another  line sinusoid",
-          color: "grey",
-          thickness: "1px",
-          type: "line",
-          id: 'mySeries1'
-        }
-      ],
-      axes: {x: {key: "x"}}
-}
